6_website-chat_rag/requirements.txt
langchain
streamlit
langchain_openai
beautifulsoup4
chromadb
python_dotenv